move `prover_result`, `context_extension` to `prover` module
and export their types in `prover` module;
