+++ ergo-lib/src/chain/transaction/unsigned.rs
@@ -1,7 +1,6 @@
 //! Unsigned (without proofs) transaction
 
-use super::input::Input;
-use super::input::UnsignedInput;
+use super::input::{Input, UnsignedInput};
 #[cfg(feature = "json")]
 use super::json;
 use super::DataInput;
@@ -13,10 +12,8 @@ use super::{
 use crate::chain::transaction::ErgoBox;
 #[cfg(feature = "json")]
 use crate::chain::transaction::TransactionFromJsonError;
-use crate::{
-    chain::prover_result::{ProofBytes, ProverResult},
-    serialization::SigmaSerializable,
-};
+use crate::serialization::SigmaSerializable;
+use crate::sigma_protocol::prover::{ProofBytes, ProverResult};
 #[cfg(feature = "json")]
 use core::convert::TryFrom;
 #[cfg(feature = "json")]
