+++ ergo-lib/src/sigma_protocol/prover/context_extension.rs
