+++ ergo-lib/src/sigma_protocol/prover/prover_result.rs
@@ -1,14 +1,15 @@
 //! ProverResult
 use std::io;
 
-use super::context_extension::ContextExtension;
-use super::{Base16DecodedBytes, Base16EncodedBytes};
+use crate::chain::{Base16DecodedBytes, Base16EncodedBytes};
 use crate::serialization::{
     sigma_byte_reader::SigmaByteRead, sigma_byte_writer::SigmaByteWrite, SerializationError,
     SigmaSerializable,
 };
 #[cfg(feature = "json")]
 use serde::{Deserialize, Serialize};
+
+use super::ContextExtension;
 
 /// Serialized proof generated by ['Prover']
 #[cfg_attr(feature = "json", derive(Serialize, Deserialize))]
